Refactor base scrape functions.
