+++ internal/sources/scrape/scrape.go
@@ -1,4 +1,4 @@
-package sources
+package scrape
 
 import (
 	"log"
@@ -7,7 +7,7 @@ import (
 	"github.com/PuerkitoBio/goquery"
 )
 
-func ReadDocument(url string) (*goquery.Document, error) {
+func GetDocumentByUrl(url string) (*goquery.Document, error) {
 	// request page
 	res, err := http.Get(url)
 	if err != nil {
